feat(affelnet): Autofill consistant mef for affelnet
