+++ server/src/common/model/index.js
@@ -32,4 +32,5 @@ module.exports = {
   AfReconciliation: createModel("afreconciliation", schema.afReconciliationSchema),
   Etablissement: createModel("etablissement", schema.etablissementSchema),
   PsFormation2021: createModel("psformations2021", schema.psFormation2021Schema),
+  SandboxFormation: createModel("sandboxformation", schema.mnaFormationSchema),
 };
