+++ server/src/jobs/common/utils/referenceUtils.js
@@ -5,6 +5,9 @@ const toBePublishedRules = [
   { intitule_long: { $ne: "" } },
   { intitule_court: { $ne: null } },
   { intitule_court: { $ne: "" } },
+  { cfd_outdated: { $ne: true } },
+  { published: true },
+  { etablissement_reference_catalogue_published: true },
 ];
 
 module.exports = { toBePublishedRules };
