+++ server/src/logic/updaters/mnaFormationUpdater.js
@@ -1,11 +1,13 @@
 const logger = require("../../common/logger");
 const Joi = require("joi");
+const { aPublierRules, aPublierSoumisAValidationRules } = require("../../jobs/pertinence/affelnet/rules");
+const { asyncForEach } = require("../../common/utils/asyncUtils");
 const { findRcoFormationFromConvertedId, getPeriodeTags } = require("../../jobs/common/utils/rcoUtils");
 const { cfdMapper } = require("../mappers/cfdMapper");
 const { codePostalMapper } = require("../mappers/codePostalMapper");
 const { etablissementsMapper } = require("../mappers/etablissementsMapper");
 const { diffFormation } = require("../common/utils/diffUtils");
-const { PendingRcoFormation } = require("../../common/model");
+const { PendingRcoFormation, SandboxFormation } = require("../../common/model");
 
 const formationSchema = Joi.object({
   cfd: Joi.string().required(),
@@ -107,9 +109,6 @@ const mnaFormationUpdater = async (formation, { withHistoryUpdate = true, withCo
       }
     }
 
-    // fill mef 10 with affelnet data if we have it
-    formation.mef_10_code = formation.mef_10_code ?? formation.affelnet_mef_10_code;
-
     const updatedFormation = {
       ...formation,
       ...cfdMapping,
@@ -120,6 +119,44 @@ const mnaFormationUpdater = async (formation, { withHistoryUpdate = true, withCo
       update_error,
       uai_formation,
     };
+
+    // try to fill mefs for Affelnet
+    // reset field value
+    updatedFormation.affelnet_mefs_10 = null;
+    if (updatedFormation.mefs_10?.length > 0) {
+      //  filter mefs10 to get affelnet_mefs_10
+      await SandboxFormation.deleteMany({});
+
+      // eslint-disable-next-line no-unused-vars
+      const { _id, ...rest } = updatedFormation;
+
+      // Split formation into N formation with 1 mef each
+      // & insert theses into a tmp collection
+      await asyncForEach(updatedFormation.mefs_10, async (mefObj) => {
+        await new SandboxFormation({
+          ...rest,
+          mef_10_code: mefObj.mef10,
+          mefs_10: [mefObj],
+        }).save();
+      });
+
+      // apply pertinence filters against the tmp collection
+      const results = await SandboxFormation.find(
+        {
+          $or: [aPublierRules, aPublierSoumisAValidationRules],
+        },
+        { mef_10_code: 1 }
+      ).lean();
+
+      if (results && results.length > 0) {
+        // keep the successful mefs in affelnet field
+        updatedFormation.affelnet_mefs_10 = results.reduce((acc, { mef_10_code }) => {
+          return [...acc, mef_10_code];
+        }, []);
+      }
+
+      await SandboxFormation.deleteMany({});
+    }
 
     const { updates, keys } = diffFormation(formation, updatedFormation);
     if (updates) {
